Adopt ICLR 2024 preprint template; stop Mac model runs
Template matches the ICLR 2024 preprint look (per the example arXiv template): single-column Times, 'Preprint' running header with rule, small-caps left-aligned title and author block, centred small-caps ABSTRACT with an indented block, small-caps numbered section/subsection headings. Removed the custom burgundy title rules.

Mac model runs abandoned (3B on MPS exhausted 25.8GB RAM and crashed the machine); all model experiments run on the H200 only. Mac is for PDF builds and lightweight analysis only.

+++ liar/tokenset.py
@@ -49,7 +49,7 @@ def statistical_token_set(
             logits = lm.model(**enc).logits
             lengths = enc["attention_mask"].sum(dim=1) - 1
             first = logits[torch.arange(len(batch)), lengths]  # (B, V)
-            acc += first.double().sum(dim=0).cpu()
+            acc += first.float().sum(dim=0).cpu().double()  # MPS has no float64
             n += len(batch)
         return (acc / n).float()
